// src/models/case.model.ts
export default class CaseModel {
  Country: string | undefined;
  CountryCode: string | undefined;
  Slug: string | undefined;
  NewConfirmed: number | undefined;
  TotalConfirmed: number | undefined;
  TotalRecovered: number | undefined;
  NewDeaths: number | undefined;
  TotalDeaths: number | undefined;
  NewRecovered: number | undefined;
  Date: string | undefined;
}